+++ malipo-flutter/lib/malipo.dart
@@ -3,5 +3,7 @@ export 'src/exceptions.dart';
 export 'src/client.dart';
 export 'src/resources/charges.dart';
 export 'src/resources/transactions.dart';
+export 'src/resources/refunds.dart';
+export 'src/resources/checkout_sessions.dart';
 export 'src/resources/balance.dart';
 export 'src/resources/webhooks.dart';